+++ src/c++/lib/applications/GetChromDepth/ChromDepthOptions.cpp
@@ -17,8 +17,6 @@
 //
 //
 
-/// \file
-
 #include "ChromDepthOptions.hh"
 
 #include "blt_util/log.hh"
@@ -58,8 +56,8 @@ parseOptions(
     ChromDepthOptions& opt,
     std::string& errorMsg)
 {
-    if (checkStandardizeInputFile(opt.alignmentFilename, "alignment", errorMsg)) return true;
-    if (checkStandardizeInputFile(opt.referenceFilename, "reference fasta", errorMsg)) return true;
+    if (checkAndStandardizeRequiredInputFilePath(opt.alignmentFilename, "alignment", errorMsg)) return true;
+    if (checkAndStandardizeRequiredInputFilePath(opt.referenceFilename, "reference fasta", errorMsg)) return true;
 
     if (vm.count("chrom"))
     {