+++ src/c++/lib/blt_util/compat_util.hh
@@ -42,9 +42,3 @@ compat_round(const double x);
 
 const char*
 compat_basename(const char* s);
-
-
-// gets canonical name of paths, but only when these refer to existing items
-// returns false on error.
-bool
-compat_realpath(std::string& path);