+++ src/c++/lib/manta/MultiJunctionUtil.hh
@@ -29,13 +29,13 @@
 #include <vector>
 
 
-/// convert independent SV candidates into multi-junction event candidates
+/// \brief Convert independent SV candidates into multi-junction event candidates
 ///
-/// given a set of un-associated single-junction SV candidates, analyze which
+/// Given a set of un-associated single-junction SV candidates, analyze which
 /// candidate junctions could potentially be treated as a single multi-junction
 /// event (such as a reciprocal translocation)
 ///
-/// right now multi-junction events are limited to pairs of (spannning) SV candidates, where
+/// Right now multi-junction events are limited to pairs of (spannning) SV candidates, where
 /// the breakends of both junctions in the pair are proximal and meeting the expected orientation
 /// pattern consistent with a reciprocal translocation.
 ///